Use more ijar, by adding "macro" attribute (default False) to rules
This allows more frequent correct use of ijar. For example, consider

A----
    |
    v
J-->M---L
        ^
        |
B--------

where A and B are Scala, J is Java, L is non-macro Scala, and M is macro
Scala. A requires full L and M, B requires ijar L, and J requires ijar M
and ijar L.

Add deps_used_whitelist for macro only jars or similar. This could be
improved with macrotracker or a custom compiler plugin.