Include transitive_proto_paths in the call to protoc
Proto rules that specify `strip_import_prefix` or `import_prefix` may
have a different import path than their workspace-relative path, and so
may not be visible on compile. This includes the transitive proto paths
for a compile target to the ScalaPB compiler invocation ensuring that
such targets can be compiled.