Catch and downgrade StackOverflowError
Zinc seems to run into this error pretty frequently, and if uncaught,
the entire worker process will exit suddenly, which can leave Bazel in a
weird state where it tries to send messages to the worker, but can't
because the process has already exited. Consider this a temporary
workaround until the underlying issues with Bazel can be fixed.