Add Scala 3 support, update to 2.13, and update Zinc
This is a rather large commit with lots of changes. The end goal of the
commit is to support Scala 3. That goal has been met, but not extensively
tested in a production codebase.

Part of this effort included upgrading the Scala 2 version in these rules
to Scala 2.13 in order to test Scala 2.13's forwards compatibility with
Scala 3.

Zinc was also upgraded, so it supports Scala 3.

Scalafmt was also upgraded, so it supports scala 3.