Introduce a use_ijar ScalaConfiguration value to support Scala 3 and 2
Using ijar on Scala 3 jars currently breaks things. Using ijar on Scala 2
does work. In order to support both of these things simultaneously, I
introduced a config value on the Scala compiler, which informs whether the
ijar or non-ijar version of the target should be used.

I changed the Scala 3 test to make sure we had Scala 3 depending on Scala 2
and vice versa.

Ultimately it would be nice to get rid of this configuration value by
fixing ijars for Scala 3.