Change to creating URLClassloaders using getPlatformClassLoader to support Java 11
Java 11 changed the extension classloader to be the platform classloader.
With this change came some changes to what is part of the bootstrap
classloader. As a result, we need to create URLClassLoaders using the
ClassLoader.getPlatformClassLoader method to get access to many of the
classes that used to be accessible through the bootstrap classloader. For
example, java.sql classes.

For more info, see this documentation https://docs.oracle.com/en/java/javase/11/migrate/index.html#JSMIG-GUID-A868D0B9-026F-4D46-B979-901834343F9E