Handle specs2 scopes by themselves properly
b9436b0 Assumed that when the test filter is just a scope, the ::
separating the scope and the (absent) test name would be escaped along
with the scope, as it is when there is a test name. However, this is not
the case. Just the scope is escaped, and the :: is not. So, replacing
everything after the starting \Q, up to and including the ::, with
nothing gets rid of the \E (which, when there's just a scope, is before
the ::) which is needed for the \Q to work properly.