Fix bug where usePersistence being false made analyses an empty map always.
This usePersistence check was added with the depsCache. It causes issues
with Tasty files being out of sync with Scala 3 compilation. Hence we are
getting rid of the check.

Removing the check here should be fine because the Dep.create call takes
in depsCache, which is an Option. If it's not set, it doesn't get written.